move android js debugger implementation to template; modify runtimeHelper to accomodate the changes
