+++ src/src/com/tns/Configuration.java
@@ -6,9 +6,9 @@ import android.app.Application;
 
 public class Configuration
 {
-	public final Application application;
 	public final ThreadScheduler threadScheduler;
 	public final Logger logger;
+	public final Debugger debugger;
 	public final String appName;
 	public final File runtimeLibPath;
 	public final File rootDir;
@@ -17,13 +17,13 @@ public class Configuration
 	public final File dexDir;
 	public final String dexThumb;
 	
-	public Configuration(Application application, ThreadScheduler threadScheduler, Logger logger, 
+	public Configuration(ThreadScheduler threadScheduler, Logger logger,  Debugger debugger,
 			String appName, File runtimeLibPath, File rootDir, File appDir, ClassLoader classLoader,
 			File dexDir, String dexThumb)
 	{
-		this.application = application;
 		this.threadScheduler = threadScheduler;
 		this.logger = logger;
+		this.debugger = debugger;
 		this.appName = appName;
 		this.runtimeLibPath = runtimeLibPath;
 		this.rootDir = rootDir;
