+++ src/src/com/tns/Debugger.java
@@ -1,0 +1,12 @@
+package com.tns;
+
+public interface Debugger
+{
+	void start();
+	
+	void disableAgent();
+	
+	void enableAgent();
+	
+	void onConnect(JsDebugger context);
+}
